run upgrade script
