+++ applications/io_demo/rose-meta/lfric-io_demo/vn3.2/rose-meta.conf
@@ -1,0 +1,91 @@
+import=lfric-driver/vn3.2
+
+[namelist:files]
+compulsory=true
+description=Configuration options for file reading and writing
+help=Configuration options for file reading and writing
+ns=namelist/Job/IO
+sort-key=Section-A03
+title=Files
+
+[namelist:files=checkpoint_stem_name]
+compulsory=true
+description=Filename stem for checkpoint files
+help=Where this app writes and reads checkpoint files.
+ns=namelist/Job/IO
+sort-key=A-007
+!string_length=filename
+type=character
+
+[namelist:files=temporal_file_path]
+compulsory=true
+description=File path for temporal reading
+ns=namelist/Job/IO
+sort-key=A-008
+!string_length=filename
+type=character
+
+[namelist:io_demo]
+compulsory=true
+description=Provides options for configuring the runtime behaviour of the IO_Demo app
+ns=namelist/io_demo
+sort-key=Section-A02
+title=IO_Demo
+
+[namelist:io_demo=benchmark_sleep_time]
+compulsory=true
+description=Number of seconds to sleep for each timestep in I/O benchmark mode
+!kind=default
+type=integer
+
+[namelist:io_demo=io_benchmark]
+compulsory=true
+description=Configure application to run as an I/O benchmarking tool
+help=Configure application to run as an I/O benchmarking tool
+!kind=default
+trigger=namelist:io_demo=benchmark_sleep_time: .true. ;
+       =namelist:io_demo=n_benchmark_fields: .true. ;
+type=logical
+
+[namelist:io_demo=multifile_io]
+compulsory=true
+description=Use multifile_io functionality
+help=This is used to turn the multifile_io functionality in the io_demo app
+    =on and off
+!kind=default
+type=logical
+
+[namelist:io_demo=n_benchmark_fields]
+compulsory=true
+description=Number of fields created in I/O benchmark
+!kind=default
+type=integer
+
+[namelist:io_demo=temporal_reading]
+compulsory=true
+description=Demonstrate temporal reading functionality
+help=This is used to turn the temporal reading functionality in the io_demo app
+    =on and off
+!kind=default
+trigger=namelist:files=temporal_file_path: .true. ;
+type=logical
+
+[namelist:multifile_io]
+compulsory=false
+duplicate=true
+!instance_key_member=filename
+
+[namelist:multifile_io=filename]
+compulsory=true
+!kind=default
+type=character
+
+[namelist:multifile_io=start_timestep]
+compulsory=true
+!kind=default
+type=integer
+
+[namelist:multifile_io=stop_timestep]
+compulsory=true
+!kind=default
+type=integer
