+++ components/lfric-xios/rose-meta/lfric-xios/vn3.2/rose-meta.conf
@@ -1,0 +1,1 @@
+import=lfric-driver/vn3.2
