+++ infrastructure/source/utilities/lfric_core_version_mod.f90
@@ -15,10 +15,10 @@ module lfric_core_version_mod
   private
   public lfric_core_version_char
 
-  integer(i_def), public, parameter  :: lfric_core_major_version = 3
-  integer(i_def), public, parameter  :: lfric_core_minor_version = 1
-  integer(i_def), public, parameter  :: lfric_core_patch_version = 1
-  logical(l_def), public, parameter  :: lfric_core_release_version = .false.
+  integer(i_def), public, parameter  :: lfric_core_major_version  = 3
+  integer(i_def), public, parameter  :: lfric_core_minor_version  = 2
+  integer(i_def), public, parameter  :: lfric_core_patch_version  = 0
+  logical(l_def), public, parameter  :: lfric_core_release_version  = .true.
   character(2), parameter            :: prefix = 'vn'
   character(4), parameter            :: dev_suffix = '_dev'
