+++ rose-stem/app/coupled/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-coupled/vn3.1_t324
+meta=lfric-coupled/vn3.2
 
 [command]
 default=rose app-run -C $CYLC_WORKFLOW_RUN_DIR/app/coupled -O lam -O LAM50x50-2x2 --install-only
