+++ rose-stem/app/io_demo/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-io_demo/vn3.1_t324
+meta=lfric-io_demo/vn3.2
 
 [command]
 default=$CORE_ROOT_DIR/bin/tweak_iodef ; \
