+++ rose-stem/app/lbc_demo/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-lbc_demo/vn3.1_t324
+meta=lfric-lbc_demo/vn3.2
 
 [command]
 default=$CORE_ROOT_DIR/bin/tweak_iodef ; \
