+++ rose-stem/app/mesh/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-mesh_tools/vn3.1_t270
+meta=lfric-mesh_tools/vn3.2
 
 [command]
 default=echo "There is no default mesh generator, please specify an optional configuration"; false
