+++ rose-stem/app/mesh_tools/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-mesh_tools/vn3.1_t270
+meta=lfric-mesh_tools/vn3.2
 
 [command]
 default=${BIN_DIR}/${APPLICATION} mesh_generation.nml
