+++ rose-stem/app/simple_diffusion/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-simple_diffusion/vn3.1_t324
+meta=lfric-simple_diffusion/vn3.2
 
 [command]
 default=$CORE_ROOT_DIR/bin/tweak_iodef ; \
