+++ rose-stem/app/skeleton/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-skeleton/vn3.1_t324
+meta=lfric-skeleton/vn3.2
 
 [command]
 default=$LAUNCH_SCRIPT/launch-exe
